+++ server/schemas/user.js
@@ -2,7 +2,7 @@ const userTypeDefs = `
   type User {
     id: Int
     username: String
-    seatingCapacity: Int
+    seatingCapacity: Float
   }
 
   extend type Query {
@@ -13,7 +13,7 @@ const userTypeDefs = `
   extend type Mutation {
     register(username: String!, password: String!): String
     login(username: String!, password: String!): String
-    updateSeatingCapacity(seatingCapacity: Int!): Int
+    updateSeatingCapacity(seatingCapacity: Float!): Float
   }
 `;
